requirements.txt
httpx==0.27.2
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
anthropic==0.40.0
slack-bolt==1.21.3
slack-sdk==3.33.4
apscheduler==3.10.4
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
pytz==2024.2
uvicorn==0.32.1
fastapi==0.115.6
